api/requirements.txt
fastapi==0.115.6
Flask==2.2.3
Flask_Cors==5.0.0
gdown==5.2.0
numpy==1.23.5
opencv_python==4.10.0.84
Pillow==11.0.0
PyYAML==6.0.2
torch==2.5.1
torchvision==0.20.1
transformers==4.46.3
uvicorn==0.32.1
Werkzeug==2.2.2

